aws-cloud-drive/CODE/requirements.txt
Flask==3.0.0
Authlib==1.3.0
boto3==1.34.0
psycopg[binary]
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
requests==2.31.0
